Move RetryActivity to the corresponding file
