+++ service/history/execution/mutable_state_builder_methods_child_workflow.go
@@ -25,10 +25,8 @@ package execution
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/uber/cadence/common"
-	"github.com/uber/cadence/common/backoff"
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/persistence"
 	"github.com/uber/cadence/common/types"
@@ -534,59 +532,4 @@ func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionTimedOutEvent(
 	initiatedID := attributes.GetInitiatedEventID()
 
 	return e.DeletePendingChildExecution(initiatedID)
-}
-
-func (e *mutableStateBuilder) RetryActivity(
-	ai *persistence.ActivityInfo,
-	failureReason string,
-	failureDetails []byte,
-) (bool, error) {
-
-	opTag := tag.WorkflowActionActivityTaskRetry
-	if err := e.checkMutability(opTag); err != nil {
-		return false, err
-	}
-
-	if !ai.HasRetryPolicy || ai.CancelRequested {
-		return false, nil
-	}
-
-	now := e.timeSource.Now()
-
-	backoffInterval := getBackoffInterval(
-		now,
-		ai.ExpirationTime,
-		ai.Attempt,
-		ai.MaximumAttempts,
-		ai.InitialInterval,
-		ai.MaximumInterval,
-		ai.BackoffCoefficient,
-		failureReason,
-		ai.NonRetriableErrors,
-	)
-	if backoffInterval == backoff.NoBackoff {
-		return false, nil
-	}
-
-	// a retry is needed, update activity info for next retry
-	ai.Version = e.GetCurrentVersion()
-	ai.Attempt++
-	ai.ScheduledTime = now.Add(backoffInterval) // update to next schedule time
-	ai.StartedID = common.EmptyEventID
-	ai.RequestID = ""
-	ai.StartedTime = time.Time{}
-	ai.TimerTaskStatus = TimerTaskStatusNone
-	ai.LastFailureReason = failureReason
-	ai.LastWorkerIdentity = ai.StartedIdentity
-	ai.LastFailureDetails = failureDetails
-
-	if err := e.taskGenerator.GenerateActivityRetryTasks(
-		ai.ScheduleID,
-	); err != nil {
-		return false, err
-	}
-
-	e.updateActivityInfos[ai.ScheduleID] = ai
-	e.syncActivityTasks[ai.ScheduleID] = struct{}{}
-	return true, nil
 }
